Automatically backed up by Learn

+++ db/migrate/001_create_costumes.rb
@@ -2,8 +2,8 @@ class CreateCostumes < ActiveRecord::Migration[5.1]
   def change
     create_table :costumes do |t|
       t.string :name
-      t.decimal :price 
-      t.decimal :size 
+      t.float :price 
+      t.float :size 
       t.string :image_url
       t.timestamp :created_at
       t.timestamp :updated_at
